initial project structure

+++ constants/http_responses.go
@@ -2,5 +2,5 @@ package constants
 
 import "net/http"
 
-const OK int16 = http.StatusOK
-const ERROR_CODE_404 int16 = http.StatusNotFound
+const OK int = http.StatusOK
+const ERROR_CODE_404 int = http.StatusNotFound
